HW1. add class Person builder
